facebook

+++ delhelp.php
@@ -1,5 +1,6 @@
 <?php 
 require_once("connect.php");
+if($_POST['action'] == 1){
  $id = $_POST["id"];
  $sql = "DELETE FROM `help` WHERE `id`= ".$id;
  echo $sql;
@@ -7,4 +8,15 @@ require_once("connect.php");
  if($result){
   echo "success";
  }else $con->error ;
+}
+
+if($_POST['action'] == 2){
+    $id = $_POST["id"];
+    $sql = "UPDATE help set  Used = 'No'  WHERE `id`= ".$id;
+    echo $sql;
+    $result = $con->query($sql);
+    if($result){
+     echo "success";
+    }else $con->error ;
+}
 ?>
